+++ CHANGELOG.md
@@ -1,3 +1,14 @@
+### Added
+- [testCaseId](https://reportportal.io/docs/Test-case-ID%3Ewhat-is-it-test-case-id) reporting via [`ReportingApi.setTestCaseId`](README.md#setTestCaseId)
+- Support nested steps for Cucumber reporting via `cucumberNestedSteps` flag
+- `skippedIssue` parameter to not mark skipped tests as 'To Investigate' by default
+- TypeScript definitions provided
+
+### Updated
+- `@reportportal/client-javascript` bumped to version `5.0.6`
+
+### Changed
+- Package size reduced
 
 ## [5.0.1] - 2021-10-05
 ### Fixed